move util functions into single file
