--- src/utils/clamp.js
@@ -1,3 +1,0 @@
-export function clamp(value, min, max) {
-  return Math.min(Math.max(value, min), max);
-}
